+++ tests/test_export.py
@@ -290,8 +290,8 @@ def test_export_definitions(tmp_empty, code, expected):
 
 
 @pytest.mark.parametrize('code, expected', [
-    [with_definitions, None],
-    [definition_with_import, 'matplotlib\n'],
+    [with_definitions, 'ploomber\n'],
+    [definition_with_import, 'ploomber\nmatplotlib\n'],
 ],
                          ids=[
                              'with_definitions',
@@ -301,12 +301,9 @@ def test_export_requirements(tmp_empty, code, expected):
     exporter = export.NotebookExporter(_read(code))
     exporter.export_requirements()
 
-    if expected is None:
-        assert not Path('requirements.txt').exists()
-    else:
-        expected = ('# Auto-generated file, may need manual '
-                    f'editing\n{expected}')
-        assert Path('requirements.txt').read_text() == expected
+    expected = ('# Auto-generated file, may need manual '
+                f'editing\n{expected}')
+    assert Path('requirements.txt').read_text() == expected
 
 
 def test_does_not_create_exported_py_if_no_definitions(tmp_empty):